+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * Fixed bug [#210](https://github.com/final-hill/decorator-contracts/issues/210) for supporting subcontract private fields
 * Added `within` feature for declaring timing requirements
 * Updated documentation
+* Third party features are subject to a class's `invariant` when apply/call is used
+* A contracted feature can not be applied to objects of a different base class
 
 ## v0.20.5